emergency_app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
torch
openai-whisper
ffmpeg-python
python-multipart==0.0.9
python-dotenv==1.0.1
cachetools==5.4.0
slowapi==0.1.9
pydantic==2.9.0
